Move all vulcan-core components into lesswrong package
