+++ packages/lesswrong/components/vulcan-core/Datatable.jsx
@@ -1,8 +1,7 @@
 import { Components, registerComponent, getCollection } from 'meteor/vulcan:lib';
+import { withCurrentUser, withMulti } from 'meteor/vulcan:core';
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import withCurrentUser from '../containers/withCurrentUser.js';
-import withMulti from '../containers/withMulti.js';
 import { FormattedMessage, intlShape } from 'meteor/vulcan:i18n';
 import { getFieldValue } from './Card.jsx';
 import _sortBy from 'lodash/sortBy';
