+++ packages/lesswrong/components/vulcan-core/vulcan-core-components.ts
@@ -1,0 +1,10 @@
+// Also imported by Storybook
+import { importComponent } from 'meteor/vulcan:lib';
+
+importComponent("App", () => require('./App'));
+importComponent("Card", () => require('./Card'));
+importComponent("Datatable", () => require('./Datatable'));
+importComponent(["EditButton", "EditForm"], () => require('./EditButton'));
+importComponent("Loading", () => require('./Loading'));
+importComponent(["NewButton", "NewForm"], () => require('./NewButton'));
+importComponent("ScrollToTop", () => require('./ScrollToTop'));
