--- packages/vulcan-core/lib/modules/components.js
@@ -1,11 +1,0 @@
-/*
-
-Also imported by Storybook
-
-*/
-import { importComponent } from 'meteor/vulcan:lib';
-
-importComponent("Card", () => require('./components/Card.jsx'));
-importComponent("Loading", () => require('./components/Loading.jsx'));
-importComponent("NewButton", () => require('./components/NewButton.jsx'));
-importComponent("EditButton", () => require('./components/EditButton.jsx'));
